requirements.txt
asgiref==3.8.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-js-asset==2.2.0
django-utils-six==2.0
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
